socketReceiveBytes: replace size and indexing ints with size_t

+++ src/Sodaq_nbIOT.cpp
@@ -648,7 +648,7 @@ size_t Sodaq_nbIOT::socketReceiveHex(char* buffer, size_t length, SaraN2UDPPacke
 
 size_t Sodaq_nbIOT::socketReceiveBytes(uint8_t* buffer, size_t length, SaraN2UDPPacketMetadata* p)
 {
-    int size = min(length, min(MAX_UDP_BUFFER, _pendingUDPBytes));
+    size_t size = min(length, min(MAX_UDP_BUFFER, _pendingUDPBytes));
 
     SaraN2UDPPacketMetadata packet;
     char tempBuffer[MAX_UDP_BUFFER];
@@ -656,7 +656,7 @@ size_t Sodaq_nbIOT::socketReceiveBytes(uint8_t* buffer, size_t length, SaraN2UDP
     size_t receivedSize = socketReceive(p ? p : &packet, tempBuffer, size);
 
     if (buffer && length > 0) {
-        for (int i = 0; i < receivedSize; i += 2) {
+        for (size_t i = 0; i < receivedSize; i += 2) {
             buffer[i / 2] = HEX_PAIR_TO_BYTE(tempBuffer[i], tempBuffer[i + 1]);
         }
     }
